+++ jans-auth-server/server/src/main/java/io/jans/as/server/authorize/ws/rs/DeviceAuthorizationAction.java
@@ -12,8 +12,8 @@ import io.jans.as.model.util.Util;
 import io.jans.as.server.i18n.LanguageBean;
 import io.jans.as.server.model.common.DeviceAuthorizationCacheControl;
 import io.jans.as.server.model.common.DeviceAuthorizationStatus;
-import io.jans.as.server.model.common.SessionId;
-import io.jans.as.server.model.common.SessionIdState;
+import io.jans.as.common.model.session.SessionId;
+import io.jans.as.common.model.session.SessionIdState;
 import io.jans.as.server.service.CookieService;
 import io.jans.as.server.service.DeviceAuthorizationService;
 import io.jans.as.server.service.SessionIdService;
