+++ jans-auth-server/server/src/main/java/io/jans/as/server/revoke/RevokeSessionRestWebService.java
@@ -10,8 +10,8 @@ import io.jans.as.common.model.common.User;
 import io.jans.as.model.common.ComponentType;
 import io.jans.as.model.error.ErrorResponseFactory;
 import io.jans.as.model.session.EndSessionErrorResponseType;
-import io.jans.as.server.model.common.SessionId;
-import io.jans.as.server.model.common.SessionIdState;
+import io.jans.as.common.model.session.SessionId;
+import io.jans.as.common.model.session.SessionIdState;
 import io.jans.as.server.model.config.Constants;
 import io.jans.as.server.model.session.SessionClient;
 import io.jans.as.server.security.Identity;
