+++ jans-auth-server/server/src/main/java/io/jans/as/server/security/Identity.java
@@ -7,7 +7,7 @@
 package io.jans.as.server.security;
 
 import io.jans.as.common.model.common.User;
-import io.jans.as.server.model.common.SessionId;
+import io.jans.as.common.model.session.SessionId;
 import io.jans.as.server.model.session.SessionClient;
 
 import jakarta.annotation.Priority;
