+++ jans-auth-server/server/src/main/java/io/jans/as/server/service/DeviceAuthorizationService.java
@@ -13,7 +13,7 @@ import io.jans.as.model.configuration.AppConfiguration;
 import io.jans.as.model.error.ErrorResponseFactory;
 import io.jans.as.server.model.common.DeviceAuthorizationCacheControl;
 import io.jans.as.server.model.common.DeviceAuthorizationStatus;
-import io.jans.as.server.model.common.SessionId;
+import io.jans.as.common.model.session.SessionId;
 import io.jans.service.CacheService;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
