+++ jans-auth-server/server/src/main/java/io/jans/as/server/service/expiration/ExpirationNotificatorTimer.java
@@ -8,7 +8,7 @@ package io.jans.as.server.service.expiration;
 
 import io.jans.as.model.config.StaticConfiguration;
 import io.jans.as.model.configuration.AppConfiguration;
-import io.jans.as.server.model.common.SessionId;
+import io.jans.as.common.model.session.SessionId;
 import io.jans.as.server.service.cdi.event.ExpirationEvent;
 import io.jans.as.server.service.external.ExternalApplicationSessionService;
 import io.jans.as.server.service.external.session.SessionEvent;
