+++ jans-auth-server/server/src/main/java/io/jans/as/server/service/external/context/ConsentGatheringContext.java
@@ -11,7 +11,7 @@ import io.jans.as.common.model.registration.Client;
 import io.jans.as.common.service.common.UserService;
 import io.jans.as.model.configuration.AppConfiguration;
 import io.jans.as.server.authorize.ws.rs.ConsentGatheringSessionService;
-import io.jans.as.server.model.common.SessionId;
+import io.jans.as.common.model.session.SessionId;
 import io.jans.jsf2.service.FacesService;
 import io.jans.model.SimpleCustomProperty;
