+++ jans-auth-server/server/src/main/java/io/jans/as/server/service/external/context/ExternalPostAuthnContext.java
@@ -7,7 +7,7 @@
 package io.jans.as.server.service.external.context;
 
 import io.jans.as.common.model.registration.Client;
-import io.jans.as.server.model.common.SessionId;
+import io.jans.as.common.model.session.SessionId;
 import io.jans.model.custom.script.conf.CustomScriptConfiguration;
 
 import jakarta.servlet.http.HttpServletRequest;
