+++ jans-auth-server/server/src/main/java/io/jans/as/server/service/external/session/SessionEvent.java
@@ -6,7 +6,7 @@
 
 package io.jans.as.server.service.external.session;
 
-import io.jans.as.server.model.common.SessionId;
+import io.jans.as.common.model.session.SessionId;
 import io.jans.model.custom.script.conf.CustomScriptConfiguration;
 
 import jakarta.servlet.http.HttpServletRequest;
