+++ jans-auth-server/server/src/main/java/io/jans/as/server/uma/authorization/UmaAuthorizationContext.java
@@ -14,7 +14,7 @@ import io.jans.as.model.configuration.AppConfiguration;
 import io.jans.as.model.uma.persistence.UmaPermission;
 import io.jans.as.model.uma.persistence.UmaResource;
 import io.jans.as.persistence.model.Scope;
-import io.jans.as.server.model.common.SessionId;
+import io.jans.as.common.model.session.SessionId;
 import io.jans.as.server.service.external.context.ExternalScriptContext;
 import io.jans.as.server.uma.service.RedirectParameters;
 import io.jans.as.server.uma.service.UmaPermissionService;
